refacto with feature playlist controller
