+++ assets/styles/_favorites.scss
@@ -17,13 +17,6 @@
     height: 40px;
 }
 
-.playlist {
-    padding: 15px;
-    width: 400px;
-    max-height: 540px;
-    overflow-y: scroll;
-}
-
 .text-song-playlist {
     padding-left: 20px;
 }
