+++ assets/styles/_playlist.scss
@@ -17,7 +17,7 @@ i.icon-playlist {
     /* Add additional styling as needed */
 }
 
-.addToPlaylistClass {
+.playlist-collection {
     min-width: 200px;
     cursor: pointer;
     &:hover {
