+++ templates/playlist/_playlist_popup.html.twig
@@ -8,7 +8,7 @@
                         <img src="{{ song.photoAlbum }}" alt="song image" class="img-fluid" width="30px" height="30px">
                     {% endfor %}
                 </div>
-                <a class="addToPlaylistClass p-3" data-playlist-id-url="{{ path('app_playlist_id', {'playlistId': element.playlist.id, 'songId': song.id, }) }}">
+                <a class="playlist-collection p-3" data-playlist-id-url="{{ path('playlist_add', {'playlistId': element.playlist.id, 'songId': song.id, }) }}">
                     {{ element.playlist.name }}
                 </a>
             </div>
